move clap struct def
