+++ nlprule/src/bin/compile.rs
@@ -3,6 +3,20 @@ use nlprule::compile::{compile, BuildOptions, Error};
 use std::io::BufWriter;
 use fs_err as fs;
 
+#[derive(clap::Clap)]
+#[clap(
+    version = env!("CARGO_PKG_VERSION"),
+    author = "Benjamin Minixhofer <bminixhofer@gmail.com>"
+)]
+pub struct BuildOptions {
+    #[clap(long, parse(from_os_str))]
+    pub build_dir: PathBuf,
+    #[clap(long, parse(from_os_str))]
+    pub tokenizer_out: PathBuf,
+    #[clap(long, parse(from_os_str))]
+    pub rules_out: PathBuf,
+}
+
 fn main() -> Result<(), Error> {
     env_logger::init();
     let opts = BuildOptions::parse();
